# Faster collect_locations and playthrough generation