--- a/Location.py
+++ b/Location.py
@@ -59,11 +59,11 @@ class Location(object):
         return (self.parent_region.can_fill(item, manual) and self.item_rule(self, item))
 
 
-    def can_reach(self, state):
+    def can_reach(self, state, noparent=False):
         if self.is_disabled():
             return False
 
-        return state.with_spot(self.access_rule, spot=self) and state.can_reach(self.parent_region)
+        return state.with_spot(self.access_rule, spot=self) and (noparent or state.can_reach(self.parent_region))
 
 
     def is_disabled(self):