--- a/OoTRandomizer.py
+++ b/OoTRandomizer.py
@@ -32,7 +32,7 @@ def start():
         sys.exit(0)
 
     # ToDo: Validate files further than mere existance
-    if not os.path.isfile(settings.rom):
+    if settings.compress_rom != 'None' and not os.path.isfile(settings.rom):
         input('Could not find valid base rom for patching at expected path %s. Please run with -h to see help for further information. \nPress Enter to exit.' % settings.rom)
         sys.exit(1)