This implements a BFS simulation of a playthrough, tracking the regions accessed in each sphere and caching those regions and the exits to new regions that can't yet be taken, rather than use the can_reach inward DFS (inward because it starts at the goal location and searches backward until it finds an accessed region). This cuts generation time down nearly 50% depending on settings, and hopefully it's easier to understand. There may be more room for the Fill algorithm to take advantage of the new Playthrough caching, but the majority of time spent is in create_playthrough so that was what I was focusing on.

Experimental times on my laptop:
- Accessible, count=10:  33s -> 19s
- Accessible, worlds=5: 2m16s -> 1m10s
- Sanity (ocarina/egg/songs/cows/scrubs/shops/tokens/keys/maps+compasses): 5.5s -> 2.8s
- Sanity x10 with custom script to skip playthrough generation: 34s -> 18s