@r0bd0g 2019-03-09
" # This ensures the playthrough shows items being collected in the proper order."
Can you explain to me why this part is no longer necessary?