@Zannick 2019-03-09
I misunderstood the intent of that section and assumed I had it covered with the other section. It looks like I did not. Fixed, thank you.

The playthroughs for my test seed are slightly different still due to the changes in iteration order within a sphere (eg. where to find Deku Sticks), but this should at least make them accurate. I don't see any differences in WotH.