@Zannick 2019-03-09
Fixed having disabled locations show up in playthroughs by readding it to the accessibility check, but I think the list of disabled locations could be cached? Certainly it'd be easy to exclude exactly the locations listed in disabled_locations, but I'm not really sure I understand the DisableType stuff.