@TestRunnerSRL 2019-03-10
So not sure if you're aware, but if you have a dictionary or set you intend to iterate over, then you need to make sure the keys properly implement hashing functions. They must have both an equals and a hash function and if two objects evaluate as equal, then the has must be equal. By default, the Obj class just uses the object's pointer for the hash/equals method. The problem is that this is not consistent for the same seeding, the ordering will change on each generation and is effectively random. There was a very similar bug with this a while back in the State class. You can try to resolve this by having implementing these functions, or by storing something that already has the implemented, like Strings. Which is why a lot of these structs have been storing lists/sets of strings instead of the actual obj they represent.

The other problem I have is that I'm not really a huge fan of how extra accessibility rules are generated. I'm worried about this becoming harder to maintain in the future. Would it be possible to call the original can_reach rules still? If you're just trying to ignore the parent reachability, maybe add some kind of parameter to ignore the parent can_reach. Something like this already exists for locations since that is how beatable_only works of the most part.