@Zannick 2019-03-11
Thanks for the feedback.

I'd prefer to keep the actual objects themselves in the sets to avoid constantly needing to map get_location, especially since I'm keeping all the worlds' locations in the same set instead of splitting them up. So I'll write some hashing functions, maybe something like hash(name + world + seed).

As for the other part, yeah, sorry, sloppy of me to use the access rules directly like that. I will add a noparent parameter to Location.can_reach and Entrance.can_reach (I don't think I want to mix State.can_reach into this).