@Zannick 2019-03-11
Ah, python's hashing semantics include hash randomization that is intentionally different on each run, so order will always be effectively random for sets. This is affecting the order of WotH locations and locations in a playthrough sphere... and hints. :( (Likely there may also be cases where it affects what's in a sphere.)

This is apparently not the case for dicts which are maintaining insertion order, so I'm going to have to replace the sets by lists and dicts and redo some of the set logic.