@Zannick 2019-03-12
I've replaced the sets that were causing the nondeterministic spoiler output, and simplified the access checks to be noparent calls to can_reach. For my worlds=5 test case, this is a pretty minor performance hit (down to 1m15s-ish) within the margin of error so I'm happy with it. I've also added MS to BDoT -> ToT which makes it required. (Otherwise, the algorithm would never need MS as it bypasses can_become_adult for speed reasons.)